fix to user reducer not loading on refresh
