+++ frontend/src/reducers/userReducer.js
@@ -4,6 +4,8 @@ export const userReducer = (state = initial_state, action) => {
   switch (action.type) {
     case "SET_USER":
       return action.payload;
+    case "REMOVE_USER":
+      return initial_state;
     default:
       return state;
   }
